requirements.txt
streamlit
pandas
matplotlib
seaborn
xlsxwriter
plotly.express
openpyxl